Update book.php
i removed the if statement checking if the submitted fields are empty. it's redundant. the 'required' attribute in the input tags already checks for that. i also changed the database connection string details. and i added the commas to the insert values for the sql statement, that's what was causing the "white page" ... jeeezzzzz there's no closing html tag !!